+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     * Spin boxes for numeric properties now allow basic math
     * Significantly improved spin box for 2d val
     * Small improvements to the gradients docks (visual changes, easier to edit, and duplicate button)
+    * Added "Focus Mode" which hides all controls from the canvas
 * I/O:
     * Improved matte import and export for lottie
     * Improved mask properties import for lottie